feat(types): move common types in domain/common & superTypes folders
