+++ src/domain/common/type.d.ts
